converted to double prec. and win64

+++ windows/stream/stream/streampower.h
@@ -23,32 +23,32 @@ public:
 
 	// parameters
 	int lattice_size_x, lattice_size_y, duration, printinterval, printstep;
-	float K, D, timestep, deltax, thresh, thresholdarea, time;
-	float xllcorner, yllcorner, nodata;
+	double K, D, timestep, deltax, thresh, thresholdarea, time;
+	double xllcorner, yllcorner, nodata;
 
 	// internal variables
 	std::vector<int> iup, idown, jup, jdown, topovecind;
-	std::vector<float> ax, ay, bx, by, cx, cy, ux, uy, rx, ry, topovec;
-	std::vector<std::vector<float>> topo, topoold, topo2, slope, flow, flow1, flow2, flow3, flow4, flow5, flow6, flow7, flow8, U;
-	Array2D<float> elevation;
+	std::vector<double> ax, ay, bx, by, cx, cy, ux, uy, rx, ry, topovec;
+	std::vector<std::vector<double>> topo, topoold, topo2, slope, flow, flow1, flow2, flow3, flow4, flow5, flow6, flow7, flow8, U;
+	Array2D<double> elevation;
 
-	static float Ran3(std::default_random_engine& generator, std::uniform_real_distribution<float>& distribution);
-	static float Gasdev(std::default_random_engine& generator, std::normal_distribution<float>& distribution);
-	static std::vector<int> Indexx(std::vector<float>& arr);	// done
-	static void Tridag(std::vector<float>& a, std::vector<float>& b, std::vector<float>& c, std::vector<float>& r, std::vector<float>& u, int n); // done
+	static double Ran3(std::default_random_engine& generator, std::uniform_real_distribution<double>& distribution);
+	static double Gasdev(std::default_random_engine& generator, std::normal_distribution<double>& distribution);
+	static std::vector<int> Indexx(std::vector<double>& arr);	// done
+	static void Tridag(std::vector<double>& a, std::vector<double>& b, std::vector<double>& c, std::vector<double>& r, std::vector<double>& u, int n); // done
 
 	StreamPower(Parameters p);
 	~StreamPower();
 
-	std::vector<std::vector<float>> CreateRandomField();
-	std::vector<std::vector<float>> ReadArcInfoASCIIGrid(char* fname); 
-	std::vector<std::vector<float>> GetTopo(); 
+	std::vector<std::vector<double>> CreateRandomField();
+	std::vector<std::vector<double>> ReadArcInfoASCIIGrid(char* fname);
+	std::vector<std::vector<double>> GetTopo();
 
 	void AssignVariables();
-	void SetInitialValues(std::vector<std::vector<float>> t);
-	void SetTopo(std::vector<std::vector<float>> t); 
-	void SetU(std::vector<std::vector<float>> u);
-	void SetU(float u);
+	void SetInitialValues(std::vector<std::vector<double>> t);
+	void SetTopo(std::vector<std::vector<double>> t);
+	void SetU(std::vector<std::vector<double>> u);
+	void SetU(double u);
 	void SetupGridNeighbors();
 	void HillSlopeDiffusion();
 	void InitDiffusion();
@@ -59,7 +59,7 @@ public:
 	void Start();
 	void Step();
 	void PrintState(char* fname);
-	void Init(int nx, int ny, float xllcorner, float yllcorner, float deltax, float nodata);
+	void Init(int nx, int ny, double xllcorner, double yllcorner, double deltax, double nodata);
 };
 
 // http://stackover_flow.com/questions/1577475/c-sorting-and-keeping-track-of-indexes
